dom/serviceworkers/test/test_serviceworker_interfaces.html

<!-- Any copyright is dedicated to the Public Domain.
   - http://creativecommons.org/publicdomain/zero/1.0/ -->
<!DOCTYPE HTML>
<html>
<head>
  <title>Validate Interfaces Exposed to Service Workers</title>
  <script src="/tests/SimpleTest/SimpleTest.js"></script>
  <link rel="stylesheet" type="text/css" href="/tests/SimpleTest/test.css" />
  <script src="../../tests/mochitest/general/interface_exposure_checker.js"></script>
  <script src="../worker_driver.js"></script>
</head>
<body>
<script class="testbody" type="text/javascript">

  function setupSW(registration) {
    var iframe;
    var worker = registration.installing ||
                 registration.waiting ||
                 registration.active;
    window.onmessage = function(event) {
      if (event.data.type == 'finish') {
        iframe.remove();
        registration.unregister().then(function(success) {
          ok(success, "The service worker should be unregistered successfully");

          SimpleTest.finish();
        }, function(e) {
          dump("Unregistering the SW failed with " + e + "\n");
          SimpleTest.finish();
        });
      } else if (event.data.type == 'status') {
        ok(event.data.status, event.data.msg);

      } else if (event.data.type == 'getPrefs') {
        let result = {};
        event.data.prefs.forEach(function(pref) {
          result[pref] = SpecialPowers.Services.prefs.getBoolPref(pref);
        });
        worker.postMessage({
          type: 'returnPrefs',
          prefs: event.data.prefs,
          result
        });

      } else if (event.data.type == 'getHelperData') {
        worker.postMessage({
          type: 'returnHelperData',
          result: getHelperData(),
        });
      }
    }

    worker.onerror = function(event) {
      ok(false, 'Worker had an error: ' + event.data);
      SimpleTest.finish();
    };

    iframe = document.createElement("iframe");
    iframe.src = "message_receiver.html";
    iframe.onload = function() {
      worker.postMessage({ script: "test_serviceworker_interfaces.js" });
    };
    document.body.appendChild(iframe);
  }

  function runTest() {
    navigator.serviceWorker.register("serviceworker_wrapper.js", {scope: "."})
      .then(setupSW);
  }

  SimpleTest.waitForExplicitFinish();
  onload = function() {
    var prefs = [
      ["dom.serviceWorkers.exemptFromPerDomainMax", true],
      ["dom.serviceWorkers.enabled", true],
    ];
    SpecialPowers.pushPrefEnv({"set": prefs}, runTest);
  };
</script>
</body>
</html>
